# docs/content/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Lumina API
  version: 1.0.0
  description: Existing internal endpoints are mounted below /api. New public endpoints use /api/v1.
servers:
  - url: http://localhost:3000
security:
  - sessionCookie: []
paths:
...
  /api/friends:
    get:
      summary: List accepted friends
      responses:
        '200':
          description: Friends returned
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/posts:
    post:
      summary: Create a post with optional media
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                content:
                  type: string
                visibility:
                  type: string
                  enum:
                    - PUBLIC
                    - COLLEGE
                    - FRIENDS
                    - PRIVATE
                media:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Post created
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/leaderboard:
    get:
      summary: Get the paginated coding leaderboard
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Leaderboard page returned
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: better-auth.session_token
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string